requirements.txt
asgiref==3.5.0
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
Faker==13.2.0
psycopg2==2.9.3
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
